# storage/docs/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "title": "HRis",
    "description": "Human Resource and Payroll System",
    "contact": {
      "email": "bertrand.kintanar@gmail.com"
    },
    "version": "1.0.0"
  },
  "host": "api.hris.dev",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Authenticates guest user by logging in.",
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "description": "Registered user's email address",
            "required": true,
            "type": "string",
            "default": "bertrand.kintanar@gmail.com"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "Registered user's password",
            "required": true,
            "type": "string",
            "default": "retardko"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "title": "data",
              "properties": {
                "token": {
                  "description": "Authenticates guest user by logging in.",
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "401": {
            "description": "Invalid credentials",
            "schema": {
              "title": "data",
              "properties": {
                "error": {
                  "description": "Authenticates guest user by logging in.",
                  "type": "string",
                  "default": "invalid_credentials"
                }
              },
              "type": "object"
            }
          },
          "500": {
            "description": "Could not create token",
            "schema": {
              "title": "data",
              "properties": {
                "error": {
                  "description": "Authenticates guest user by logging in.",
                  "type": "string",
                  "default": "could_not_create_token"
                }
              },
              "type": "object"
            }
          }
        }
      }
    },
    "/logout": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logout currently authenticated user.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "JWT Token",
            "required": true,
            "type": "string",
            "items": {
              "type": "string"
            },
            "default": "Bearer "
          }
        ],
        "responses": {
          "200": {
            "description": "Signed out",
            "schema": {
              "title": "data",
              "properties": {
                "message": {
                  "description": "Logout currently authenticated user.",
                  "type": "string",
                  "default": "signed_out"
                },
                "status_code": {
                  "description": "Logout currently authenticated user.",
                  "type": "integer",
                  "default": 200
                }
              },
              "type": "object"
            }
          },
          "400": {
            "description": "Token not provided",
            "schema": {
              "title": "data",
              "properties": {
                "message": {
                  "description": "Logout currently authenticated user.",
                  "type": "string",
                  "default": "Token not provided"
                },
                "status_code": {
                  "description": "Logout currently authenticated user.",
                  "type": "integer",
                  "default": 400
                },
                "debug": {
                  "description": "Logout currently authenticated user.",
                  "type": "object"
                }
              },
              "type": "object"
            }
          },
          "417": {
            "description": "Cannot sign out",
            "schema": {
              "title": "data",
              "properties": {
                "message": {
                  "description": "Logout currently authenticated user.",
                  "type": "string",
                  "default": "cannot_sign_out"
                },
                "status_code": {
                  "description": "Logout currently authenticated user.",
                  "type": "integer",
                  "default": 417
                }
              },
              "type": "object"
            }
          },
          "500": {
            "description": "Could not create token",
            "schema": {
              "title": "data",
              "properties": {
                "message": {
                  "description": "Logout currently authenticated user.",
                  "type": "string",
                  "default": "could_not_create_token"
                },
                "status_code": {
                  "description": "Logout currently authenticated user.",
                  "type": "integer",
                  "default": 500
                }
              },
              "type": "object"
            }
          }
        }
      }
    },
...
  },
...
}